Work in progress: Rename and Reorder columns NOT yet fully implemented.
This will require some additional work for command parsing (Rename). This
will be very similar to the changes made for the Reorder command.

Also requires some additional work for Reorder: Need an extract function
to move data to a new CSVMatrix object, and then an insert function to
insert the contents of a CSVMatrix object into an existing CSVMatrix
object.